requirements.txt
flask
requests
beautifulsoup4
pandas
groq
python-dotenv
lxml
gunicorn
